requirements.txt
telethon>=1.35
jinja2>=2.11.3
PyYAML>=5.4.1
cryptg>=0.4
Pillow>=8.3.2
feedgen>=0.9.0
python-magic>=0.4.24
pytz>=2020.5
